code refactor and cleanup
