+++ submodules/ContactListUI/Sources/HelperClasses/LJBaseTableViewCell.swift
@@ -1,0 +1,39 @@
+//
+//  LJBaseTableViewCell.swift
+//  _idx_ContactListUI_1D7887AF_ios_min13.0
+//
+//  Created by Apple on 15/09/22.
+//
+
+import Foundation
+import UIKit
+
+class LJBaseTableViewCell: UITableViewCell {
+    
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        initView()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        initView()
+    }
+    
+    private func initView() {
+        selectionStyle = .none
+        backgroundColor = .clear
+    }
+    
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+        // Configure the view for the selected state
+    }
+
+}
