+++ submodules/ContactListUI/Sources/Search/WEVSearchRecordManager.swift
@@ -45,11 +45,7 @@ struct WEVSearchRecordManager {
             LJUserDefaults.liveSearchRecord = data
         }
     }
-    
-
 }
-
-
 
 struct LJUserDefaults {
