+++ submodules/ContactListUI/Sources/WEVRootViewController.swift
@@ -20,19 +20,9 @@ import StickerResources
 import ContextUI
 import QrCodeUI
 import ContactsUI
-//import SnapKit
-import HandyJSON
-import Alamofire
 import Supabase
 import PostgREST
-import Alamofire
 import Realtime
-//import FirebaseAuth
-//import FirebaseAnalytics
-//import FirebaseCore
-
-
-
 
 public class WEVRootViewController: ViewController {
     private let context: AccountContext
@@ -40,8 +30,8 @@ public class WEVRootViewController: ViewController {
         return context
     }
     
-    private var contactsNode: WEVRootNode {
-        return self.displayNode as! WEVRootNode
+    private var contactsNode: WEVDiscoverRootNode {
+        return self.displayNode as! WEVDiscoverRootNode
     }
     private var validLayout: ContainerViewLayout?
     
@@ -224,7 +214,7 @@ public class WEVRootViewController: ViewController {
     }
     
     override public func loadDisplayNode() {
-        self.displayNode = WEVRootNode(context: self.context, sortOrder: sortOrderPromise.get() |> distinctUntilChanged, present: { [weak self] c, a in
+        self.displayNode = WEVDiscoverRootNode(context: self.context, sortOrder: sortOrderPromise.get() |> distinctUntilChanged, present: { [weak self] c, a in
             self?.present(c, in: .window(.root), with: a)
         }, controller: self)
